processing wallet transaction body
